+++ apps/web/src/app/layout.css.ts
@@ -46,33 +46,3 @@ export const boxStyle = recipe({
     margin: 'normal',
   },
 });
-
-export const textStyle = recipe({
-  base: {
-    color: 'white',
-  },
-  variants: {
-    color: {
-      black: {
-        color: 'black',
-      },
-      white: {
-        color: 'white',
-      },
-    },
-    typography: {
-      t1: {
-        fontSize: '12px',
-        fontWeight: 300,
-      },
-      t2: {
-        fontSize: '14px',
-        fontWeight: 500,
-      },
-    },
-  },
-  defaultVariants: {
-    color: 'white',
-    typography: 't1',
-  },
-});
