+++ apps/web/src/app/layout.tsx
@@ -24,6 +24,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
+      <link rel="icon" href="/favicon.ico" sizes="any" />
       <body className={`${geistSans.variable} ${geistMono.variable}`}>{children}</body>
     </html>
   );
