--- apps/web/src/app/pagestyle.css.ts
@@ -1,7 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const pageStyle = style({
-  width: '500px',
-  height: '500px',
-  backgroundColor: 'purple',
-});
