+++ apps/web/src/assets/bottles.png
