+++ apps/web/src/components/info/infoStyle.css.ts
